First commit
This is the landing page showing a login system. Users will be able to register (pending roughly 75% current backend completion), or use default dummy creds to access the system and view tickets. Additional build-out in progress to create the ticketing view. At this time, it is not apt to work on sessions for the demo due to complexity from logistical and security perspectives- if you make it, make it stronger than your bank, right?